+++ tfac/dataImport.py
@@ -13,7 +13,7 @@ from tensorly.tenalg.svd import randomized_svd
 from .cmtf import perform_CMTF
 
 PATH_HERE = dirname(dirname(abspath(__file__)))
-OPTIMAL_SCALING = 2 ** 5.0
+OPTIMAL_SCALING = 2 ** 8.0
 
 
 @lru_cache
@@ -160,7 +160,6 @@ def form_tensor(variance_scaling: float = OPTIMAL_SCALING):
     return np.copy(tensor), np.copy(rna), patient_data
 
 
-@lru_cache
 def get_factors(variance_scaling: float = OPTIMAL_SCALING, r=8):
     """
     Return the factorization results.
@@ -175,45 +174,8 @@ def get_factors(variance_scaling: float = OPTIMAL_SCALING, r=8):
     """
     tensor, rna, patient_data = form_tensor(variance_scaling)
     np.random.seed(42)
-    t_fac = perform_CMTF(tensor, rna, r=r, maxiter=1000)
-    return t_fac, patient_data
-
-
-class PCArand(PCA):
-    def _compute_eig(self):
-        """
-        Override slower SVD methods
-        """
-        _, s, v = randomized_svd(self.transformed_data, self._ncomp)
-
-        self.eigenvals = s ** 2.0
-        self.eigenvecs = v.T
-
-
-def get_pca_factors(variance_scaling: float = OPTIMAL_SCALING, r=8):
-    """
-    Returns factorization results using PCA.
-
-    Parameters:
-        variance_scaling (float, default:1.0): RNA/cytokine variance scaling
-        r (int, default:8): PCA components to use
-
-    Returns:
-        components (numpy.array): PCA components
-        patient_data (pandas.DataFrame): patient data, including status, data
-            types, and cohort
-        var_explained (float): variance explained by PCA
-    """
-    tensor, rna, patient_data = form_tensor(variance_scaling)
-
-    unfolded_tensor = tl.unfold(tensor, 0)
-    stacked = np.hstack((unfolded_tensor, rna))
-
-    pca = PCA(stacked, ncomp=r, missing='fill-em', method='nipals')
-    components = pca.factors / abs(pca.factors).max(axis=0)
-    var_explained = pca.rsquare[-1]
-
-    return components, patient_data, var_explained
+    t_fac, pcaFac = perform_CMTF(tensor, rna, r=r)
+    return t_fac, pcaFac, patient_data
 
 
 def reorder_table(df):
