+++ tfac/figures/figure2.py
@@ -5,11 +5,9 @@ import numpy as np
 import pandas as pd
 
 from .common import getSetup
-from ..dataImport import form_tensor, get_factors, get_pca_factors
+from ..dataImport import form_tensor, get_factors
 from ..predict import run_model
 from tensorpack import calcR2X
-
-from cProfile import Profile
 
 
 def get_r2x_results():
@@ -24,12 +22,10 @@ def get_r2x_results():
         r2x_v_components (pandas.Series): R2X vs. number of CMTF components
         r2x_v_scaling (pandas.Series): R2X vs. RNA/cytokine scaling
     """
-    pf = Profile()
-    pf.enable()
     # R2X v. Components
     tensor, matrix, patient_data = form_tensor()
     labels = patient_data.loc[:, 'status']
-    components = 5
+    components = 12
 
     r2x_v_components = pd.DataFrame(
         columns=['CMTF', 'PCA'],
@@ -43,18 +39,13 @@ def get_r2x_results():
     )
     for n_components in r2x_v_components.index:
         print(f"Starting decomposition with {n_components} components.")
-        t_fac, _ = get_factors(r=n_components)
-        pca_components, _, pca_var = get_pca_factors(r=n_components)
+        t_fac, pcaFac, _ = get_factors(r=n_components)
         r2x_v_components.loc[n_components, 'CMTF'] = t_fac.R2X
-        r2x_v_components.loc[n_components, 'PCA'] = pca_var
+        r2x_v_components.loc[n_components, 'PCA'] = pcaFac.rsquare[-1]
         acc_v_components.loc[n_components, 'CMTF'] = \
             run_model(t_fac.factors[0], labels)[0]
         acc_v_components.loc[n_components, 'PCA'] = \
-            run_model(pca_components, labels)[0]
-        
-    pf.disable()
-    pf.dump_stats("profile")
-    return
+            run_model(pcaFac.scores, labels)[0]
 
     # R2X v. Scaling
     scalingV = np.logspace(-10, 10, base=2, num=21)
@@ -69,15 +60,14 @@ def get_r2x_results():
     )
     for scaling in r2x_v_scaling.index:
         tensor, matrix, _ = form_tensor(scaling)
-        t_fac, _ = get_factors(variance_scaling=scaling)
-        pca_components, _, _ = get_pca_factors(r=n_components)
+        t_fac, pcaFac, _ = get_factors(variance_scaling=scaling)
         r2x_v_scaling.loc[scaling, "Total"] = t_fac.R2X
         r2x_v_scaling.loc[scaling, "Tensor"] = calcR2X(t_fac, tIn=tensor)
         r2x_v_scaling.loc[scaling, "Matrix"] = calcR2X(t_fac, mIn=matrix)
         acc_v_scaling.loc[scaling, 'CMTF'] = \
             run_model(t_fac.factors[0], labels)[0]
         acc_v_scaling.loc[scaling, 'PCA'] = \
-            run_model(pca_components, labels)[0]
+            run_model(pcaFac.scores, labels)[0]
 
     return r2x_v_components, acc_v_components, r2x_v_scaling, acc_v_scaling
