+++ tfac/tests/test_factor.py
@@ -8,5 +8,5 @@ from ..cmtf import perform_CMTF
 def test_CMTF():
     """ Test that we can form the missing tensor. """
     tensor, matrix, _ = form_tensor()
-    tFac = perform_CMTF(tensor, matrix, r=9)
+    tFac, _ = perform_CMTF(tensor, matrix, r=8)
     assert tFac.R2X > 0.0
